fix(desktop): clear max-lines debt from git review v2 merge
- extract shared gap-expansion flattening out of Chat/SplitDiffViewer
  into hooks/ui/diff/diff-gap-flatten.ts (the two copies were identical
  modulo row shape); ratchet ChatDiffViewer allowlist 631 -> 550 and
  drop SplitDiffViewer's entry (466 now)
- extract GitReviewFileSectionShell from GitReviewFileRow (555 -> 387)
- drop the stale commit split button from the git review playground
  mock (507 -> 472), matching the shipped pane
- allowlist generated cloud SDK types at 776 (+numstat fields)
